Flask Yolo Streamlit App/requirements.txt
torch
torchvision
ultralytics
streamlit
opencv-python-headless
pillow
